services: slugify before DB lookup.
In `ldap_register()`, LDAP binding has already been performed,
and the user is allowed to authenticate. However, the current
logic is to look up the user entry in local DB using the string
from the login form. If the latter fails, creating a new user
is attempted - which might have been already created!

This still does not allow special characters (as in #43, which
can be unsafe), but does address the underlying issue.

Should fix #15, #17, #45, and make #43 obsolete.